+++ modules/core/obliqueMap.js
@@ -4,6 +4,8 @@ import {defaults as olDefaultInteractions} from "ol/interaction.js";
 import {ViewDirection, viewDirectionNames} from "vcs-oblique/src/vcs/oblique/viewDirection";
 import {transformFromImage} from "vcs-oblique/src/vcs/oblique/helpers";
 import LoaderOverlay from "../../src/utils/loaderOverlay";
+import {Radio} from "backbone";
+import store from "../../src/app-store";
 
 /**
  * Creates an oblique map on activate and handles deactivation. Handles interactions, click-events and the activation of oblique layer.
@@ -13,7 +15,8 @@ const ObliqueMap = Backbone.Model.extend({
         active: false
     },
     initialize: function () {
-        const channel = Radio.channel("ObliqueMap");
+        const channel = Radio.channel("ObliqueMap"),
+            obliqueLayers = Radio.request("ModelList", "getModelsByAttributes", {typ: "Oblique"});
 
         this.pausedInteractions = [];
 
@@ -77,6 +80,9 @@ const ObliqueMap = Backbone.Model.extend({
                     this.setCenter(coordinate, resolution);
                 }
             }
+        });
+        obliqueLayers.forEach(layer => {
+            this.registerLayer(layer);
         });
     },
     /**
@@ -200,6 +206,7 @@ const ObliqueMap = Backbone.Model.extend({
 
         if (this.isActive() && this.currentCollection && this.currentDirection?.currentImage) {
             Radio.trigger("Map", "beforeChange", "2D");
+            store.commit("Map/setMapMode", "2D");
             this.deactivateOpenTreeOnTopicSearch();
             this.getCenter().then(function (center) {
                 const resolutionFactor = this.currentLayer.get("resolution"),
@@ -400,6 +407,7 @@ const ObliqueMap = Backbone.Model.extend({
 
             this.setActiveToolToFalse();
             map2D = Radio.request("Map", "getMap");
+            store.commit("Map/setMapMode", "Oblique");
 
             if (!this.container) {
                 this.container = this.createAndInsertTargetContainer(map2D);
@@ -438,6 +446,7 @@ const ObliqueMap = Backbone.Model.extend({
                     this.activateLayer(layer, center, resolution).then(function () {
                         layer.set("isVisibleInMap", true);
                         layer.set("isSelected", true);
+                        // here the switching off of all not-oblique layers is triggered
                         Radio.trigger("Map", "change", "Oblique");
                         LoaderOverlay.hide();
                     });